requirements.txt
fireworks_ai==0.15.3
openai==1.60.2
Requests==2.32.3
tqdm==4.66.5
